LayoutTests/http/tests/media/media-stream/disconnected-frame-permission-denied-expected.txt

Tests that when a getUserMedia request is made, permission is denied and its Frame is disconnected before a callback is made, the error callback is invoked with PERMISSION_DENIED.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.name is "NotAllowedError"

PASS Error callback invoked
PASS successfullyParsed is true

TEST COMPLETE
